+++ dashboard/app/controllers/openai_personalization_controller.rb
@@ -1,0 +1,24 @@
+require 'json'
+
+class OpenaiPersonalizationController < ApplicationController
+  authorize_resource class: false
+  # Need to modify this... modify abilities.rb?
+  skip_authorization_check only: [:match_teaching_profile]
+
+  # POST /openai/match_teaching_profile
+  def match_teaching_profile
+    teaching_profile_data = match_teaching_profile_params[:teaching_profile_data]
+
+    unless teaching_profile_data
+      return render status: :bad_request, json: {error: "Missing teaching_profile_data"}
+    end
+
+    response = OpenaiPersonalizationHelper.match_teaching_profile(teaching_profile_data)
+
+    return render(status: response[:status], json: response[:json])
+  end
+
+  private def match_teaching_profile_params
+    params.permit(:teaching_profile_data => {})
+  end
+end
